Modularise Cordova tasks
