+++ gulp/cordova/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+var gulp = require('gulp');
+
+gulp.task('cordova', ['cordova-build', 'cordova-move']);
